+++ tests/fixtures/project-monorepo/app/dirty.py
@@ -1,0 +1,8 @@
+def in_scope_one():
+    # TODO: implement
+    return None
+
+
+def in_scope_two():
+    # FIXME: handle errors
+    pass